+++ Material.Blazor/Components/Grid/MBEnumerationsGrid.cs
@@ -1,0 +1,22 @@
+// ToDo:
+//
+//  Move enumerations to MBEnumerations
+//  
+
+namespace Material.Blazor
+{
+    public enum MB_Grid_Measurement
+    {
+        EM,
+        FitToData,
+        Percent,
+        PX,
+    }
+
+    public enum MB_Grid_ColumnType
+    {
+        Icon,
+        Text,
+        TextColor,
+    };
+}